refactor(DeleteAccount): decompose DeleteAccountForm into smaller, reusable components and custom hooks following SOLID principles
