+++ src/components/Dashboard/SettingsSection/SettingsContent/DeleteAccountForm/DeleteAccountForm.component.tsx
@@ -1,82 +1,35 @@
-import { memo, useState, useCallback } from "react";
-import { motion } from "framer-motion";
-import { Button } from "../../../../UI/Button/Button.component";
-import { useDeleteAccountForm } from "../../hooks/useDeleteAccountForm";
-import { WarningBox } from "../../components/DeleteAccount/WarningBox/WarningBox.component";
-import { ConfirmationForm } from "../../components/DeleteAccount/ConfirmationForm/ConfirmationForm.component";
-import { useToast } from "../../contexts/ToastContext";
-import { AccountError } from "../../utils/api/account";
+import { memo } from "react";
+import { useConfirmationState } from "../../hooks/useConfirmationState";
+import { useAccountDeletion } from "../../hooks/useAccountDeletion";
+import { InitialWarning } from "./components/InitialWarning/InitialWarning.component";
+import { ConfirmationForm } from "./components/ConfirmationForm/ConfirmationForm.component";
 
 export const DeleteAccountForm = memo(() => {
-  const [showConfirmation, setShowConfirmation] = useState(false);
-  const { showToast } = useToast();
-  
-  const { form, onSubmit, isDeleting } = useDeleteAccountForm({
-    onSuccess: () => showToast('Konto zostało usunięte', 'success'),
-    onError: (error) => {
-      if (error instanceof AccountError) {
-        switch (error.code) {
-          case 'INVALID_PASSWORD':
-            showToast('Podane hasło jest nieprawidłowe', 'error');
-            return;
-          case 'INVALID_CONFIRMATION':
-            showToast('Nieprawidłowe potwierdzenie', 'error');
-            return;
-        }
-      }
-      showToast('Nie udało się usunąć konta', 'error');
-    }
-  });
-  
-  const handleCancel = useCallback(() => {
-    try {
-      setShowConfirmation(false);
-      form.reset();
-      showToast('Operacja została anulowana', 'success');
-    } catch (error) {
-      showToast('Nie udało się anulować operacji', 'error');
-    }
-  }, [form, showToast]);
+  const { 
+    showConfirmation, 
+    handleShowConfirmation, 
+    handleHideConfirmation 
+  } = useConfirmationState();
 
-  const handleShowConfirmation = useCallback(() => {
-    setShowConfirmation(true);
-  }, []);
+  const { 
+    form, 
+    onSubmit, 
+    isDeleting, 
+    handleCancel 
+  } = useAccountDeletion(handleHideConfirmation);
 
   if (!showConfirmation) {
-    return (
-      <motion.div
-        initial={{ opacity: 0 }}
-        animate={{ opacity: 1 }}
-        className="w-full max-w-md sm:ml-0 sm:mr-auto px-4 sm:px-0"
-      >
-        <WarningBox />
-        <Button
-          type="button"
-          onClick={handleShowConfirmation}
-          className="w-full px-6 py-3 rounded-lg bg-red-500/20 text-red-400 
-            hover:bg-red-500/30 transition-colors duration-200 shadow-none
-            text-base sm:text-lg font-medium"
-        >
-          Chcę usunąć konto
-        </Button>
-      </motion.div>
-    );
+    return <InitialWarning onConfirm={handleShowConfirmation} />;
   }
 
   return (
-    <motion.div 
-      initial={{ opacity: 0 }}
-      animate={{ opacity: 1 }}
-      className="w-full max-w-md sm:ml-0 sm:mr-auto px-4 sm:px-0"
-    >
-      <ConfirmationForm
-        register={form.register}
-        errors={form.formState.errors}
-        isSubmitting={isDeleting}
-        onCancel={handleCancel}
-        onSubmit={onSubmit}
-      />
-    </motion.div>
+    <ConfirmationForm
+      register={form.register}
+      errors={form.formState.errors}
+      isSubmitting={isDeleting}
+      onCancel={handleCancel}
+      onSubmit={onSubmit}
+    />
   );
 });
