+++ src/components/Dashboard/SettingsSection/SettingsContent/DeleteAccountForm/components/WarningBox/data/WarningBox.data.ts
