+++ src/components/Dashboard/SettingsSection/SettingsContent/DeleteAccountForm/components/WarningBox/style/WarningBox.styles.ts
