+++ src/components/Dashboard/SettingsSection/hooks/useConfirmationState.ts
@@ -1,0 +1,19 @@
+import { useState, useCallback } from 'react';
+
+export const useConfirmationState = () => {
+  const [showConfirmation, setShowConfirmation] = useState(false);
+
+  const handleShowConfirmation = useCallback(() => {
+    setShowConfirmation(true);
+  }, []);
+
+  const handleHideConfirmation = useCallback(() => {
+    setShowConfirmation(false);
+  }, []);
+
+  return {
+    showConfirmation,
+    handleShowConfirmation,
+    handleHideConfirmation
+  };
+};
